+++ packages/orq-rc/src/lib/config.ts
@@ -68,8 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "ix-sse-errors-for-speakeasy-compatibility",
+  sdkVersion: "3.12.0-rc.11",
   genVersion: "2.687.1",
-  userAgent:
-    "speakeasy-sdk/typescript ix-sse-errors-for-speakeasy-compatibility 2.687.1 2.0 @orq-ai/node",
+  userAgent: "speakeasy-sdk/typescript 3.12.0-rc.11 2.687.1 2.0 @orq-ai/node",
 } as const;